DAO
DAO

+++ app/src/main/java/com/trutechinnovations/zeus/Song.java
@@ -13,14 +13,14 @@ import java.net.URL;
 public class Song {
 
     private String name;
-    private int time;
+    private float time;
     private String artist;
     private String url;
     private ImageView imageView;
     private String imageUrl;
     private DownloadImageTask task;
 
-    public Song(String name, int time, String artist, String songUrl, String imageUrl){
+    public Song(String name, float time, String artist, String songUrl, String imageUrl){
         this.name = name;
         this.time = time;
         this.artist = artist;
@@ -43,11 +43,11 @@ public class Song {
         return name;
     }
 
-    public int getTime() {
+    public float getTime() {
         return time;
     }
 
-    public void setTime(int time) {
+    public void setTime(float time) {
         this.time = time;
     }
